pyproject.toml
[tool.poetry]
name = "flat"
version = "0.1.0"
description = ""
authors = ["Aliaksei Sofin <sofin.moffin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.73.0"
SQLAlchemy = "1.4.31"
gunicorn = "20.1.0"
uvicorn = "0.17.4"
alembic = "1.7.6"
psycopg2-binary = "2.9.3"
typer = "0.4.0"
email-validator = "1.1.3"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
SQLAlchemy-Utils = "^0.38.2"
spacy = "^3.2.2"
python-telegram-bot = "^13.11"
requests = "^2.27.1"
trafilatura = "^1.0.0"
rq = "^1.10.1"

[tool.poetry.dev-dependencies]
black = "22.1.0"
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
